backends/student_mgr/src/requirements.txt
pydantic
pydantic-settings
pydantic_yaml
fastapi
pandas
uvicorn
pyyaml